fixes  - unedited strings.xml (moved to metrolist_strings.xml)  - rename SleepTimer.start() _type param to timerType  - pluralised sleep_timer_default_set_songs string  - removed unused/unworking logic for fadeout  - set the volume multiplier back to 1  - clear unused imports  - ensure sleepTimerDefaultType is valid  - clear sleep timer songs on playlist end  - stringified hardcoded strings  - fixed log message  - extracted sleep timer dialog to a new component
