+++ app/src/main/res/values/metrolist_strings.xml
@@ -249,8 +249,19 @@
     <!-- Sleep timer -->
     <string name="set_as_default">Set as default</string>
     <string name="already_set_as_default">Already set as default</string>
+    <string name="sleep_timer_time">Time</string>
+    <string name="sleep_timer_time_fs">Time (FS)</string>
+    <string name="sleep_timer_time_fs_desc">ⓘ FS: Finishes Song before stopping</string>
+    <string name="sleep_timer_songs">Songs</string>
     <string name="sleep_timer_default_set">Sleep timer default set to %d min</string>
-    <string name="sleep_timer_default_set_songs">Sleep timer default set to %d songs</string>
+    <plurals name="sleep_timer_default_set_songs">
+        <item quantity="one">Sleep timer default set to 1 song</item>
+        <item quantity="other">Sleep timer default set to %d songs</item>
+    </plurals>
+    <plurals name="song">
+        <item quantity="one">1 song</item>
+        <item quantity="other">%d songs</item>
+    </plurals>
     <string name="enable_automatic_sleeptimer">Enable automatic sleep timer</string>
     <string name="sleeptimer_description">Enables the sleep timer automatically with the default value by a custom time</string>
     <string name="sleep_timer_repeat_description">Set a custom day and time when the sleep timer should automatically activate</string>
