+++ app/src/main/res/values/strings.xml
@@ -208,10 +208,6 @@
         <item quantity="one">1 minute</item>
         <item quantity="other">%d minutes</item>
     </plurals>
-    <plurals name="song">
-        <item quantity="one">1 song</item>
-        <item quantity="other">%d songs"</item>
-    </plurals>
     <string name="error_no_stream">No stream available</string>
     <string name="error_no_internet">No network connection</string>
     <string name="error_timeout">Timeout</string>
